refactor: extract generic JWT validator from GitHub-specific implementation
